+++ routes/leader-router.js
@@ -12,7 +12,13 @@ leaderRouter.route("/").get((req, res, next)=>{
 		res.json(leaders);
 	},(err)=>next(err))
 	.catch((err)=> next(err));
-}).post(authenticate.verifyUser,(req, res, next)=> {
+}).post(authenticate.verifyUser, (req, res, next)=> {
+	if(!authenticate.verifyAdmin(req.user)){
+		res.statusCode=403;
+		res.end('Unauthorized : need admin rights');
+		return;
+	}
+	
 	Leaders.create(req.body)
 	.then((leader)=>{
 		res.statusCode= 200;
@@ -20,11 +26,17 @@ leaderRouter.route("/").get((req, res, next)=>{
 		res.json(leader);
 	},(err)=> next(err))
 	.catch((err) => next(err));
-}).put(authenticate.verifyUser ,(req, res, next)=> {
+}).put(authenticate.verifyUser ,  (req, res, next)=> {
 	res.statusCode=405;
 	res.end('PUT not supported on /leaders');
 	
-}).delete(authenticate.verifyUser,(req, res, next)=> {
+}).delete(authenticate.verifyUser, (req, res, next)=> {
+	if(!authenticate.verifyAdmin(req.user)){
+		res.statusCode=403;
+		res.end('Unauthorized : need admin rights');
+		return;
+	}
+	
 	Leaders.remove({})
 	.then((resp)=>{
 		res.statusCode= 200;
@@ -42,10 +54,17 @@ leaderRouter.route("/:leaderId").get((req, res, next)=>{
 		res.json(leader);
 	},(err)=> next(err))
 	.catch((err)=> next(err));
-}).post(authenticate.verifyUser,(req, res, next)=> {
+}).post(authenticate.verifyUser, (req, res, next)=> {
+
 	res.statusCode=405;
 	res.end('POST not supported on /leaders/'+req.params.leaderId);
-}).put(authenticate.verifyUser,(req, res, next)=> {
+}).put(authenticate.verifyUser,  (req, res, next)=> {
+	if(!authenticate.verifyAdmin(req.user)){
+		res.statusCode=403;
+		res.end('Unauthorized : need admin rights');
+		return;
+	}
+	
 	Leaders.findByIdAndUpdate(req.params.leaderId, {$set: req.body},{new: true})
 	.then((leader)=>{
 		res.statusCode= 200;
@@ -54,7 +73,13 @@ leaderRouter.route("/:leaderId").get((req, res, next)=>{
 	},(err)=>next(err))
 	.catch((err)=> next(err));
 	
-}).delete(authenticate.verifyUser,(req, res, next)=> {
+}).delete(authenticate.verifyUser, (req, res, next)=> {
+	if(!authenticate.verifyAdmin(req.user)){
+		res.statusCode=403;
+		res.end('Unauthorized : need admin rights');
+		return;
+	}
+	
 	Leaders.findByIdAndRemove(req.params.leaderId)
 	.then((resp)=>{
 		res.statusCode= 200;
